Move artemis request object to separate class
