+++ src/test/k6/util/utils.js
